Relocate all-in-one test
