+++ src/test/groovy/com/paiondata/aristotle/service/CommonServiceSpec.java
