+++ src/test/groovy/com/paiondata/aristotle/service/GraphServiceSpec.java
