+++ src/test/groovy/com/paiondata/aristotle/service/NodeServiceSpec.java
